requirements.txt
openai-whisper>=20230314
torch>=2.0.0
numpy>=1.20.0
